# POC: speed up remote builds by caching `GetActionResult` and `FindMissingBlobs`