before:

```
INFO: Elapsed time: 94.825s, Critical Path: 51.39s
INFO: 11014 processes: 5504 remote cache hit, 5510 internal.
INFO: Build completed successfully, 11014 total actions
```

after:

```
INFO: Elapsed time: 67.897s, Critical Path: 26.42s
INFO: 11014 processes: 5504 remote cache hit, 5510 internal.
INFO: Build completed successfully, 11014 total actions
```